test(core): move the guide half of the adjudication contract to the roster
`packs/core/tests/skills/new-spec/test_acceptance_criteria_discipline.py`
reached above its own pack to assert on `guides/core/`, which
`pack-tests-stay-in-pack` rejects. It failed `make build-check` and the
caps enforcer self-test on three lines.

Move `test_planning_guide_explains_spec_review_triage` and
`test_core_explanation_places_adjudication_before_repair` to
`tests/roster/test_spec_review_adjudication_documentation.py`, where
repository-level coverage belongs, and drop `REPO_ROOT` from the pack
test. The assertions are unchanged; only their owning suite moves.
