+++ packs/core/tests/skills/new-spec/test_acceptance_criteria_discipline.py
@@ -6,13 +6,10 @@ from pathlib import Path
 import pytest
 
 PACK_ROOT = Path(__file__).resolve().parents[3]
-REPO_ROOT = PACK_ROOT.parents[1]
 SKILL = PACK_ROOT / ".apm/skills/new-spec/SKILL.md"
 SPEC = PACK_ROOT / ".apm/skills/new-spec/assets/spec.md"
 PLAN = PACK_ROOT / ".apm/skills/new-spec/assets/plan.md"
 EVALS = PACK_ROOT / ".apm/skills/new-spec/evals/evals.json"
-PLANNING_GUIDE = REPO_ROOT / "guides/core/how-to/plan-and-execute-non-trivial-work.md"
-CORE_EXPLANATION = REPO_ROOT / "guides/core/explanation/core-pack.md"
 
 SOURCES = {"skill": SKILL, "spec": SPEC, "plan": PLAN}
 RULES = (
@@ -303,25 +300,3 @@ def test_spec_review_triage_eval_has_required_shape_and_behaviour() -> None:
     assert any("blind spot" in assertion.lower() for assertion in entry["assertions"])
     assert any("validated path" in assertion.lower() for assertion in entry["assertions"])
     assert any("pre-execute" in assertion.lower() for assertion in entry["assertions"])
-
-
-def test_planning_guide_explains_spec_review_triage() -> None:
-    body = flattened(PLANNING_GUIDE)
-    assert (
-        "Every completed `adversarial-reviewer` report, including a clean claim, "
-        "goes through `finding-adjudicator`"
-    ) in body
-    assert "Only sustained findings can change the spec or plan" in body
-    assert "`draft-origin` or `prior-round-repair`" in body
-    assert "unresolved origin stops for your direction" in body
-    assert "what that gate proves and one relevant blind spot" in body
-
-
-def test_core_explanation_places_adjudication_before_repair() -> None:
-    body = flattened(CORE_EXPLANATION)
-    assert "**`finding-adjudicator`**" in body
-    report = "Every completed spec-review report, including a clean claim"
-    gateway = "passes through `finding-adjudicator` before it can change the spec or plan"
-    assert report in body
-    assert gateway in body
-    assert body.index(report) < body.index(gateway)
